Refactor split out yaep_tree.c
