+++ src/main/c/xmq.c
@@ -52,6 +52,7 @@ WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 #include"parts/yaep_util.h"
 #include"parts/yaep_symbols.h"
 #include"parts/yaep_terminal_bitset.h"
+#include"parts/yaep_tree.h"
 
 // XMQ STRUCTURES ////////////////////////////////////////////////
 
@@ -4966,6 +4967,7 @@ char *xmqLinePrintf(XMQLineConfig *lc, const char *element_name, ...)
 #include"parts/yaep_util.c"
 #include"parts/yaep_symbols.c"
 #include"parts/yaep_terminal_bitset.c"
+#include"parts/yaep_tree.c"
 #include"parts/yaep.c"
 
 #endif
